build: clear warnings

+++ minmea.h
@@ -249,7 +249,7 @@ static inline float minmea_tocoord(struct minmea_float *f)
         return NAN;
     int_least32_t degrees = f->value / (f->scale * 100);
     int_least32_t minutes = f->value % (f->scale * 100);
-    return (float) degrees + (float) minutes / (60 * f->scale);
+    return (float) degrees + (float) minutes / (float)(60 * f->scale);
 }
 
 #ifdef __cplusplus
